ml_service/requirements.txt
pika
mysql-connector-python
python-dotenv
geopandas
rasterio
requests
Pillow
torch
scipy
torchvision
mit_semseg
pyyaml
numpy
tqdm
easydict
scikit-image
scikit-learn
opencv-python
tensorflow
joblib
matplotlib
pandas
albumentations
hydra-core
pytorch-lightning
tabulate
kornia
webdataset
packaging
scikit-learn
wldhx.yadisk-direct
contextily
uvicorn
fastapi
setproctitle
boto3